// src/generated/models/Afb_ubosModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Afb_ubosafb_statutdevalidation = {
  0: 'Valid_',
  1: 'Encours',
  2: 'Nonv_rifi_',
  747010001: 'Rejet_'
} as const;
export type Afb_ubosafb_statutdevalidation = keyof typeof Afb_ubosafb_statutdevalidation;
export const Afb_ubosafb_statutppe = {
  0: 'Non',
  1: 'Auto_d_clar_e',
  747010001: 'D_tect_eparsceening'
} as const;
export type Afb_ubosafb_statutppe = keyof typeof Afb_ubosafb_statutppe;
export const Afb_ubosafb_typedentite = {
  0: 'Personnemorale',
  1: 'Personnephysique'
} as const;
export type Afb_ubosafb_typedentite = keyof typeof Afb_ubosafb_typedentite;
export const Afb_ubosstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Afb_ubosstatecode = keyof typeof Afb_ubosstatecode;
export const Afb_ubosstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Afb_ubosstatuscode = keyof typeof Afb_ubosstatuscode;

export interface Afb_ubosBase {
  afb_datedenaissance?: string;
  afb_datedevalidation?: string;
  "afb_dernierresultatdescreening@odata.bind"?: string;
  afb_nationalite?: string;
  afb_nomouraisonsociale: string;
  afb_paysderesidencefiscale?: string;
  "afb_piecedidentite@odata.bind"?: string;
  afb_pourcentagededetentiondirecte: number;
  afb_pourcentagededetentionindirecte?: number;
  afb_statutdevalidation: Afb_ubosafb_statutdevalidation;
  afb_statutppe: Afb_ubosafb_statutppe;
  "afb_tiersparent@odata.bind"?: string;
  afb_typedentite: Afb_ubosafb_typedentite;
  afb_uboid: string;
  "afb_uboparent@odata.bind"?: string;
  "afb_validepar@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Afb_ubosstatecode;
  statuscode?: Afb_ubosstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Afb_ubos extends Afb_ubosBase {
  afb_dernierresultatdescreeningname?: string;
  afb_piecedidentitename?: string;
  afb_statutdevalidationname?: string;
  afb_statutppename?: string;
  afb_tiersparentname?: string;
  afb_typedentitename?: string;
  afb_uboparentname?: string;
  afb_valideparname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  afb_dernierresultatdescreening?: object;
  _afb_dernierresultatdescreening_value?: string;
  afb_piecedidentite?: object;
  _afb_piecedidentite_value?: string;
  afb_tiersparent?: object;
  _afb_tiersparent_value?: string;
  afb_uboparent?: object;
  _afb_uboparent_value?: string;
  afb_validepar?: object;
  _afb_validepar_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}